refactor: move command api implementations
