+++ src/cli.ts
@@ -11,12 +11,12 @@ program
   .name('vivliostyle')
   .version(version, '-v, --version')
   .command('init', 'create vivliostyle config', {
-    executableFile: 'commands/init',
+    executableFile: 'init',
   })
   .command('build', 'build and create PDF file', {
-    executableFile: 'commands/build',
+    executableFile: 'build',
   })
   .command('preview', 'launch preview server', {
-    executableFile: 'commands/preview',
+    executableFile: 'preview',
   })
   .parse(process.argv);
